requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
paho-mqtt==1.6.1
requests==2.31.0
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3 